Set jvm args up
